change num_tokens to int to fix a sign comparison issue

+++ module/owlib/src/c/jsmn.c
@@ -16,7 +16,7 @@ $Id$
  * Allocates a fresh unused token from the token pull.
  */
 static jsmntok_t *jsmn_alloc_token(jsmn_parser *parser, 
-		jsmntok_t *tokens, size_t num_tokens) {
+		jsmntok_t *tokens, int num_tokens) {
 	jsmntok_t *tok;
 	if (parser->toknext >= num_tokens) {
 		return NULL;
@@ -42,7 +42,7 @@ static void jsmn_fill_token(jsmntok_t *token, jsmntype_t type,
  * Fills next available token with JSON primitive.
  */
 static jsmnerr_t jsmn_parse_primitive(jsmn_parser *parser, const char *js,
-		jsmntok_t *tokens, size_t num_tokens) {
+		jsmntok_t *tokens, int num_tokens) {
 	jsmntok_t *token;
 	int start;
 
@@ -77,7 +77,7 @@ found:
  * Fills next token with JSON string.
  */
 static jsmnerr_t jsmn_parse_string(jsmn_parser *parser, const char *js,
-		jsmntok_t *tokens, size_t num_tokens) {
+		jsmntok_t *tokens, int num_tokens) {
 	jsmntok_t *token;
 
 	int start = parser->pos;
@@ -126,7 +126,7 @@ static jsmnerr_t jsmn_parse_string(jsmn_parser *parser, const char *js,
  * Parse JSON string and fill tokens.
  */
 jsmnerr_t jsmn_parse(jsmn_parser *parser, const char *js, jsmntok_t *tokens, 
-		unsigned int num_tokens) {
+		int num_tokens) {
 	jsmnerr_t r;
 	int i;
 	jsmntok_t *token;
